+++ common/src/importers/keeperImporters/keeperCsvImporter.ts
@@ -1,9 +1,9 @@
-import { BaseImporter } from "./baseImporter";
-import { Importer } from "./importer";
+import { BaseImporter } from "../baseImporter";
+import { Importer } from "../importer";
 
-import { ImportResult } from "../models/domain/importResult";
+import { ImportResult } from "../../models/domain/importResult";
 
-import { FolderView } from "../models/view/folderView";
+import { FolderView } from "../../models/view/folderView";
 
 export class KeeperCsvImporter extends BaseImporter implements Importer {
   parse(data: string): Promise<ImportResult> {
